Movie-Recommender/movide_recommender/requirements.txt
tokenizers==0.21.1
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.6
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
elastic-transport==8.17.1
elasticsearch==8.13.0
et_xmlfile==2.0.0
filelock==3.18.0
fsspec==2025.5.1
idna==3.10
joblib==1.5.1
jwt==1.4.0
lz4==4.4.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pyarrow==20.0.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sql==2022.4.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
thrift==0.20.0
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
gunicorn==23.0.0
django-prometheus==2.4.0



